fix(Slider): Parent and get an float value for a floatSlider

+++ UMI3D-BROWSER/Assets/Runtime/Slider/Public/01-Scripts/SliderFloatParameterModel.cs
@@ -50,7 +50,7 @@ namespace umi3d.browserRuntime.ui.slider
 
         private void ValueUpdated(Notification notification)
         {
-            if (!notification.TryGetInfoT(SliderNotifiactionKeys.SliderUpdated.Value, out int value))
+            if (!notification.TryGetInfoT(SliderNotifiactionKeys.SliderUpdated.Value, out float value))
                 return;
 
             dto.value = value;
